Use the physical layer and improve the output.
The physical layer will miss certain patterns, but is an order of
magnitude faster at scanning.  The main amount of time spent in scanning
Intel spaces is counting through every page in the address space (32, 40
or 64 bits), not the actual scanning.  There's no real way around this
if you want to ensure you get every chunk of virtual memory.  Since the
scanner could be stopped after its first hit, this might still be
preferable, but should not be the default (particularly for an automagic
scan).