Fix up interface versioning issues.
Unfortunately one interface change was missed, getting the versions
slightly out of alignment.  Due to the 2.1 schema already being
published, the best course of action was to move to version 4.0.

This means that there's a tiny chance some schemas exist somewhere in
the world with a non-existant 3.0.0 schema.  Volatility will fail with
an error saying they're not supported.  5:S  Sorry!