Fix an issue we hit 1 in 1000 times.
Turns out, that because the name consumption is fairly lax (often lots
of excess space after the name, etc) we were reading 2 too many bytes
(because we weren't taken the length of the leaf off the available
length).  This would error when it so happened that a structure was at
the end of the type list and we'd read past the end of the tpi stream.