Remove object validate method.
With agreement from @attrc and @imhlv2, I'm removing the useless
validate method from the object.  It defaulted to returning inaccurate
information (validity undetermined, but validate returned True as if
validation had happened) and is better handled by individual classes
implementing is_valid (which seems the preferred nomenclature) and
throwing exceptions if it's called on a class that doesn't have it.