Symbols: Resolve symbols returned from symbol_space.get_symbol
This ensures that ReferenceTemplates are not returned from a call to
symbol_space.get_symbol.  This is in keeping with the get_type method.

The symbol_space containing all symbol tables has visibility of all
tables and can cross-resolve amongst them.  Asking a specific table
for a type may result in a ReferenceTemplate.  This feels like a
better solution than having individual tables ask their parent space
to fill in the details of sibling tables.

Fixes #279.