Maldump: Initial implementation
There were two options here: either use --dump and make malfind depend
on vaddump (which is incosistent because vaddump should be vadinfo
--dump in that case) or maintain consistency and not burden malfind with
additional requirements.

I went the second route, but I'll be happy to move things to a --dump
flag if that's the prefered route, as long as we apply it everywhere we
support dumping files.