Windows: Ensure we set the kernel_virtual_offset before other reqs
It appears in issue #524 that the step to fulfil symbol requirements
can throw an exception (which then prevents the kernel_virtual_offset
(which is optional) from being set appropriately.

This sets the kvo first and adds a check to where it's set.  It doesn't
get to the root of why an exception is thrown, but should ensure it's
easier to spot if it goes wrong.

Fixes #524.