Windows: Improve PDB scanning regex
The regular expression for finding PDB signatures didn't have an
apprioriate flag to treat newline characters like normal characters,
meaning that if a newline character occurred between the RSDS header
and the name of the pdb file, the pdb signature would be missed.

Also updated the filenames so that they're escaped, meaning it must be
an actual dot for the extension rather than any character.

Fixes #577