Renderers: Make appending rows massively more efficient
Previously we were generating the list of children (for most likely the
root node) for every single append statement, during which we were
recalculating the length of the list, twice.  In plugins that output a
lot of rows this would add an enourmous overhead (that likely grew as
the length of the output grew).  Without this overhead, the time taken
for the TreeGrid._append method went from 1230.0s to 2.4s.