Layers: Better checks on PAE page tables
This checks that the very top level table points to the next four pages,
as we'd expected in general. This relies on the same assumptions as the
existing PAE detection did, ie that the PAE page_map maps the next four
pages immediately.

Previously we didn't check that the top page was valid, once we found
the self-referential pointer.  This adds in an appropriate check to
reduce false positives.

Closes #631.