Initial canonical helper addition
The intel 64-bit 4-page paging mechanism allows for 48-bit virtual
addresses.  They introduced a convention that the higher bits must
be set a particular way to avoid operating system developers abusing
those bits and creating problems that would be difficult to resolve in
the future.

Volatility requires that addresses for mapping or translation fit within
the available bounds of the virtual address space.  This unfortunately
means that addresses that have the protections against abuse in place
can may live outside this range.

This provides two function (canonicalize and decanonicalize) which will
either set the appropriate sign extension or remove it.  The
decanonicalize function will return an adress outside of the address
range if the original value was not canonical.