Windows: Add `--name` option to pslist plugin
Add the capability to filter by name the windows.pslist output, where
users can define the name of the process they would like to filter out
in the process list.

The behavior of the filter is described below:

- Filter is case-sensitive, e.g. Input `--name Foo` will not return a
  process named foo.
- Filters full match only, e.g Input `--name foo`: return all processes named foo, it would not
  return foobaz.
- Input `--name foo bar`: return all processes named foo and
  bar.

Described below the semantics when `--pid` filter is defined with `--name`:

- Input `--name foo --pid 123`: return all processes named foo
  and all processes with pid 123, if these intersect, meaning the
  process 123 is also named foo, it will return only that one process.
  If these cases dont exist return the empty list.
- Input `--name foo bar --pid 123 456`: return all processes
  named foo and bar and all processes with pid 123 and 456. The
  semantics are the same as previous point but for multiple names and pids.

All cases mentioned above were tested manually.