Windows: Add filtering by type to handles.
This change adds the '--types' option to the windows.handles plugin,
whereby a user may specify one or more types of handles to filter the
output by.  Any handles not of the type(s) specified are not returned.
Specified types are treated as case-insensitive.

Because the object type index map is generated dynamically,
user-inputted types are not validated against the map of possible object
types.  For example, a user-inputted type of value "Foo" does not raise
an error and will return no results.