Added context arguments to windows.vadyarascan.
This change updated the windows.vadyarascan plugin to accept --size and
--reverse arguments, which control the overall size of the hex dump of
the YARA match and the negative offset respectively.  Additionally, the
Value column was replaced with Hexdump.

This more closely mirrors the functionality of volatility2.  If the
offset specified by --reverse is large enough to cause a read on a page
not present in the system, a warning is logged and padded data will be
returned.