Core: Support hidden by default columns in plugins
This adds the ability to provide a third argument in the column
definition ("name", "type", "extra").  Extra columns will not be shown
by the CLI by default, but can be seen by passing an empty list to the
CLI (ie, provide no parameters to the `--columns` flag).  Otherwise any
columns passed to `--columns` will be displayed and no other columns
will.  If no valid column name are passed to columns, then the normal
list of columns (not including extra) will be displayed.

All other UIs will need to implement support for the extra column,
otherwise they will silently ignore it.