Windows: Bug fixes and additions to modules and modscan
This pull request adds several changes to the modules and modscan
plugins that fix bugs, unify options and operations, and allow for
filtering by name and base address in both plugins. Previously,
modscan and modules had a substantial amount of partially duplicated
code, which led to bugs, such as:
	Only one allowed filtering by --name
	Only one cycled through the session layers to find the one hosting the kernel module
With the new inheritance and combined implementation, the correct
and complete implementation applies to both.

This PR also addresses the feedback in
https://github.com/volatilityfoundation/volatility3/pull/1099. In this PR,
--name and –base apply to the processing of modules and determine
both which plugins are displayed in output as well as which are
processed for extraction. This fixes the issue of using --filters, which
only filters the final command line output, but leads to every module
being extracted.