Linux: Add lsblk plugin
This pull request adds a plugin that mimics the linux lsblk utility. It
loops through the block devices on the machine and prints out the device
name, major, minor, removability, if the device is read only, type and
mountpoint.

Tested on linux versions 3.13-6.8 with Ubuntu, Rhel, Centos, Debian, sles, and
Oracle Linux distros. Worked with the latest version for each distro.