Remove the old .style.yapf
This was used before we shifted to black permanently.  It allowed
additional spacing between keyword arguments and their default values,
but it was inconsisnent and could get in a state where it would flip
between two weights, so every "fix" would then be "fixed" back again.

No longer feasible for a project of this size.