Core: Scanning only splits up the layer
Scanning had been rigged to chop up a layer's readable bits into the
chunks that make it up from the lower layer.  Unfortunately it did this
on chunks in the lower layer, rather than chunks readable in the higher
layer.  This meant chunks in Intel were already read in page sized
blocks.  This vastly simplifies the reading of data and brings
non-linear layer searching in sync with linear layer searching (because
there's no good reason they should be different).