Packaging: Refactor JSON data handling
This changes the way that JSON resources required for plugins are
handled within the volatility3 package. Instead of operating directly on
paths, it uses Python's builtin `importlib.resources` module to identify
a data directory with JSON files as a resource location. This greatly
simplifies how resources are found within the plugins.

References:

https://docs.python.org/3/library/importlib.resources.html#module-importlib.resources