Linux Utilities: Improve `get_kset_modules` signature
Returning a `dict` leads to an all-or-nothing result, where you either
get the complete set of modules, or an exception is encountered and
raised and you get nothing. Using an iterator allows for as many
readable modules to be yielded as possible, and the caller can still
catch any `InvalidAddressException` that gets thrown during iteration.